acceleratorlab/requirements.txt
streamlit>=1.33.0
pandas>=2.2.0
matplotlib>=3.8.0
pyyaml>=6.0.1
requests>=2.31.0
PyQt6>=6.6.0
